feat(tui): wrap leftover host marketplace and settings copy with ui()
Keep Host diagnostics, Zod descriptions, and marketplace errors on the same AST gate as the rest of the i18n stack.
