+++ lib/index.js
@@ -1,6 +1,7 @@
-import { a as languageSelection, c as saveLanguage, d as ui, f as uiLocale, l as setUiLocale, o as localeFromSettings, s as localeSettings, t as TUI_STARTUP_SERVICE, u as translateUiText } from "./startup-DznulVSj.js";
+import { a as saveLanguage, c as ui, i as localeSettings, l as uiLocale, o as setUiLocale, r as localeFromSettings, s as translateUiText, t as languageSelection } from "./locale-DbTu1PjP.js";
 import { r as DSH_COMPATIBILITY, s as dshCompatibilityError, t as measureStartup, u as launcherPrefersEnglish } from "./startup-trace-QEGhP-_i.js";
-import { n as assertCredentialFreeUrl, t as PluginMarketplace } from "./plugin-marketplace-D8-O8gcm.js";
+import { n as assertCredentialFreeUrl, t as PluginMarketplace } from "./plugin-marketplace-swCofv4S.js";
+import { t as TUI_STARTUP_SERVICE } from "./startup-Dk_6a6_Y.js";
 import { createRequire } from "node:module";
 import { InProcessApiClient, toFetchHandler } from "@deepseek-ai/dsh-host-apiproxy";
 import { randomUUID } from "node:crypto";
@@ -31738,30 +31739,30 @@ const CustomThemeSchema = z$1.object({
 	tokenColors: z$1.array(TextMateRuleSchema).max(MAX_TEXTMATE_RULES).default([])
 });
 const AppearanceSettingsSchema = z$1.object({
-	theme: z$1.string().pattern(/^(?:dark|light|custom:[a-z0-9](?:[a-z0-9-]{0,46}[a-z0-9])?)$/u).default(DEFAULT_TUI_THEME).description("SeekTTY 当前使用的界面主题。"),
-	codeTheme: z$1.string().pattern(/^(?:auto|dark|light|custom:[a-z0-9](?:[a-z0-9-]{0,46}[a-z0-9])?)$/u).default(DEFAULT_TUI_CODE_THEME).description("代码块独立主题；auto 跟随当前界面主题。"),
-	customThemes: z$1.array(CustomThemeSchema).max(MAX_CUSTOM_THEMES).default([]).description("SeekTTY 命名自定义主题。")
+	theme: z$1.string().pattern(/^(?:dark|light|custom:[a-z0-9](?:[a-z0-9-]{0,46}[a-z0-9])?)$/u).default(DEFAULT_TUI_THEME).description(ui("SeekTTY 当前使用的界面主题。", "The interface theme currently used by SeekTTY.")),
+	codeTheme: z$1.string().pattern(/^(?:auto|dark|light|custom:[a-z0-9](?:[a-z0-9-]{0,46}[a-z0-9])?)$/u).default(DEFAULT_TUI_CODE_THEME).description(ui("代码块独立主题；auto 跟随当前界面主题。", "Independent code-block theme; auto follows the current interface theme.")),
+	customThemes: z$1.array(CustomThemeSchema).max(MAX_CUSTOM_THEMES).default([]).description(ui("SeekTTY 命名自定义主题。", "Named custom SeekTTY themes."))
 });
 const BehaviorSettingsSchema = z$1.object({
 	toolCards: z$1.union([
 		"collapsed",
 		"expanded",
 		"hidden"
-	]).default(DEFAULT_TUI_BEHAVIOR.toolCards).description("工具卡片默认形态；启动时应用到当前会话。"),
-	showReasoning: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.showReasoning).description("推理内容默认是否显示。"),
-	desktopNotifications: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.desktopNotifications).description("回合完成或待审批时发送终端桌面通知。"),
-	followTerminalTitle: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.followTerminalTitle).description("终端标题跟随当前会话运行状态。"),
-	composerHistoryLimit: z$1.natural().max(MAX_COMPOSER_HISTORY).default(DEFAULT_TUI_BEHAVIOR.composerHistoryLimit).description("输入历史持久化条数；0 表示关闭。"),
-	statusElapsed: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.statusElapsed).description("状态栏显示当前回合实时耗时。"),
+	]).default(DEFAULT_TUI_BEHAVIOR.toolCards).description(ui("工具卡片默认形态；启动时应用到当前会话。", "Default tool-card shape; applied to the current session at startup.")),
+	showReasoning: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.showReasoning).description(ui("推理内容默认是否显示。", "Whether reasoning is shown by default.")),
+	desktopNotifications: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.desktopNotifications).description(ui("回合完成或待审批时发送终端桌面通知。", "Send a desktop notification when a turn completes or an approval is pending.")),
+	followTerminalTitle: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.followTerminalTitle).description(ui("终端标题跟随当前会话运行状态。", "Follow the current session status in the terminal title.")),
+	composerHistoryLimit: z$1.natural().max(MAX_COMPOSER_HISTORY).default(DEFAULT_TUI_BEHAVIOR.composerHistoryLimit).description(ui("输入历史持久化条数；0 表示关闭。", "Number of persisted composer history entries; 0 disables history.")),
+	statusElapsed: z$1.boolean().default(DEFAULT_TUI_BEHAVIOR.statusElapsed).description(ui("状态栏显示当前回合实时耗时。", "Show live elapsed time for the current turn in the status bar.")),
 	clipboardFallback: z$1.union([
 		"auto",
 		"osc52",
 		"off"
-	]).default(DEFAULT_TUI_BEHAVIOR.clipboardFallback).description("OSC 52 失败后的剪贴板回退命令；auto 按平台探测。"),
-	toolOutputLineLimit: z$1.natural().max(MAX_TOOL_OUTPUT_LINE_LIMIT).default(DEFAULT_TUI_BEHAVIOR.toolOutputLineLimit).description("展开态工具输出单块行数上限；0 表示不折叠。"),
-	diffContextLines: z$1.natural().max(MAX_DIFF_CONTEXT_LINES).default(DEFAULT_TUI_BEHAVIOR.diffContextLines).description("Diff 上下文行数。"),
-	dangerConfirmDefault: z$1.union(["cancel", "confirm"]).default(DEFAULT_TUI_BEHAVIOR.dangerConfirmDefault).description("危险确认默认焦点；cancel 表示回车不执行。"),
-	keyBindings: z$1.dict(z$1.string()).default({}).description("覆盖默认快捷键；键为绑定 id，值为 Ctrl+P 这类组合。空对象表示使用默认键位。")
+	]).default(DEFAULT_TUI_BEHAVIOR.clipboardFallback).description(ui("OSC 52 失败后的剪贴板回退命令；auto 按平台探测。", "Clipboard fallback after OSC 52 fails; auto probes the platform.")),
+	toolOutputLineLimit: z$1.natural().max(MAX_TOOL_OUTPUT_LINE_LIMIT).default(DEFAULT_TUI_BEHAVIOR.toolOutputLineLimit).description(ui("展开态工具输出单块行数上限；0 表示不折叠。", "Line cap for one expanded tool-output block; 0 means no folding.")),
+	diffContextLines: z$1.natural().max(MAX_DIFF_CONTEXT_LINES).default(DEFAULT_TUI_BEHAVIOR.diffContextLines).description(ui("Diff 上下文行数。", "Number of diff context lines.")),
+	dangerConfirmDefault: z$1.union(["cancel", "confirm"]).default(DEFAULT_TUI_BEHAVIOR.dangerConfirmDefault).description(ui("危险确认默认焦点；cancel 表示回车不执行。", "Default focus for danger confirmation; cancel means Enter does not proceed.")),
+	keyBindings: z$1.dict(z$1.string()).default({}).description(ui("覆盖默认快捷键；键为绑定 id，值为 Ctrl+P 这类组合。空对象表示使用默认键位。", "Override default shortcuts; keys are binding ids and values are chords such as Ctrl+P. An empty object uses the defaults."))
 });
 function settingsDocument(descriptor) {
 	return {
@@ -31788,7 +31789,7 @@ function createSettingsDescribeCache(load) {
 	};
 	const one = (namespace) => {
 		const document = all().find((row) => row.namespace === namespace);
-		if (document === void 0) throw new Error(`设置命名空间 ${JSON.stringify(namespace)} 已卸载`);
+		if (document === void 0) throw new Error(ui(`设置命名空间 ${JSON.stringify(namespace)} 已卸载`, `Settings namespace ${JSON.stringify(namespace)} is no longer available`));
 		return document;
 	};
 	return {
@@ -31814,9 +31815,9 @@ function sessionExportFilename(sessionId) {
 	return `dsh-session-${sessionId.replace(/[^A-Za-z0-9._-]/gu, "_").slice(0, 120) || "session"}.zip`;
 }
 function validateCatalogSource(source) {
-	if (source.builtIn || source.kind !== "catalog") throw new Error("内置 npm Source 不能写入用户来源");
-	if (!/^[a-z][a-z0-9-]*$/.test(source.id) || source.id === NPM_SOURCE.id) throw new Error(`Catalog Source id ${JSON.stringify(source.id)} 必须是唯一的小写 kebab-case`);
-	if (source.label.trim() === "" || source.url.trim() === "") throw new Error("Catalog Source 名称和 URL 不能为空");
+	if (source.builtIn || source.kind !== "catalog") throw new Error(ui("内置 npm Source 不能写入用户来源", "The built-in npm Source cannot be written as a user source"));
+	if (!/^[a-z][a-z0-9-]*$/.test(source.id) || source.id === NPM_SOURCE.id) throw new Error(ui(`Catalog Source id ${JSON.stringify(source.id)} 必须是唯一的小写 kebab-case`, `Catalog Source id ${JSON.stringify(source.id)} must be a unique lowercase kebab-case value`));
+	if (source.label.trim() === "" || source.url.trim() === "") throw new Error(ui("Catalog Source 名称和 URL 不能为空", "Catalog Source name and URL cannot be empty"));
 	assertCredentialFreeUrl(source.url, "Catalog Source URL");
 	if (source.credentialRef !== void 0 && source.credentialRef !== "") credentialRef(source.credentialRef);
 	return {
@@ -31835,7 +31836,7 @@ function degradedCatalogSource(raw, index, diagnostic) {
 	return {
 		id,
 		kind: "catalog",
-		label: `${label}（无效）`,
+		label: ui(`${label}（无效）`, `${label} (invalid)`),
 		url,
 		enabled: false,
 		builtIn: false,
@@ -31854,7 +31855,7 @@ function catalogSourcesFromStored(stored, reservedIds) {
 		const record = typeof raw === "object" && raw !== null ? raw : {};
 		const rawId = typeof record.id === "string" ? record.id : "";
 		if (rawId !== "" && seen.has(rawId)) {
-			sources.push(degradedCatalogSource(raw, index, `Catalog Source ${rawId} 与内置或 Provider Source 冲突`));
+			sources.push(degradedCatalogSource(raw, index, ui(`Catalog Source ${rawId} 与内置或 Provider Source 冲突`, `Catalog Source ${rawId} conflicts with a built-in or Provider Source`)));
 			continue;
 		}
 		try {
@@ -31868,7 +31869,7 @@ function catalogSourcesFromStored(stored, reservedIds) {
 				builtIn: false
 			});
 			if (seen.has(storedSource.id)) {
-				sources.push(degradedCatalogSource(raw, index, `Catalog Source ${storedSource.id} 与内置或 Provider Source 冲突`));
+				sources.push(degradedCatalogSource(raw, index, ui(`Catalog Source ${storedSource.id} 与内置或 Provider Source 冲突`, `Catalog Source ${storedSource.id} conflicts with a built-in or Provider Source`)));
 				continue;
 			}
 			seen.add(storedSource.id);
@@ -31892,7 +31893,7 @@ function tuiProfile(summary) {
 	return {
 		...summary,
 		compatible: false,
-		diagnostic: "Profile 未组合 TUI Surface；可以复制为新的 TUI Profile，但不能由 deepseek 直接启动"
+		diagnostic: ui("Profile 未组合 TUI Surface；可以复制为新的 TUI Profile，但不能由 deepseek 直接启动", "This Profile does not compose a TUI Surface; copy it into a new TUI Profile instead of launching it with deepseek")
 	};
 }
 async function mutateSettings(settings, namespace, ops, expectedRevision) {
@@ -31915,7 +31916,7 @@ function createTuiManagementBridge(ctx, cwd) {
 	const providers = ctx.get("tuiMarketplaceProviders");
 	const settings = ctx.settings;
 	const credentials = ctx.credentials;
-	if (manager === void 0) throw new Error("tui-runner: Settings、Credentials 或 Profile Plugin Manager 未装配");
+	if (manager === void 0) throw new Error(ui("tui-runner: Settings、Credentials 或 Profile Plugin Manager 未装配", "tui-runner: Settings, Credentials, or the Profile Plugin Manager is not mounted"));
 	settings.register(MARKETPLACE_NAMESPACE, MarketplaceSettingsSchema, { applies: "live" });
 	settings.register(APPEARANCE_NAMESPACE, AppearanceSettingsSchema, { applies: "live" });
 	settings.register(BEHAVIOR_NAMESPACE, BehaviorSettingsSchema, { applies: "live" });
@@ -31941,7 +31942,7 @@ function createTuiManagementBridge(ctx, cwd) {
 	return {
 		sessionExport: { download: async (sessionId, includeDescendants, signal) => {
 			const apiProxy = ctx.get("apiProxy");
-			if (apiProxy === void 0) throw new Error("Harness Session Export 服务未装配");
+			if (apiProxy === void 0) throw new Error(ui("Harness Session Export 服务未装配", "Harness Session Export service is not mounted"));
 			const request = {
 				sessionId,
 				...includeDescendants ? { includeDescendants: true } : {}
@@ -31949,9 +31950,9 @@ function createTuiManagementBridge(ctx, cwd) {
 			const response = await apiProxy.downloads.sessionLog(request, signal ?? new AbortController().signal);
 			if (!response.ok) {
 				const detail = (await response.text()).trim().slice(0, 1e3);
-				throw new Error(`Harness Session Export 失败（HTTP ${String(response.status)}）${detail === "" ? "" : `：${detail}`}`);
+				throw new Error(ui(`Harness Session Export 失败（HTTP ${String(response.status)}）${detail === "" ? "" : `：${detail}`}`, `Harness Session Export failed (HTTP ${String(response.status)})${detail === "" ? "" : `: ${detail}`}`));
 			}
-			if (response.body === null) throw new Error("Harness Session Export 返回了空响应体");
+			if (response.body === null) throw new Error(ui("Harness Session Export 返回了空响应体", "Harness Session Export returned an empty response body"));
 			const rawLength = response.headers.get("content-length");
 			const contentLength = rawLength === null ? void 0 : Number.parseInt(rawLength, 10);
 			return {
@@ -32011,10 +32012,10 @@ function createTuiManagementBridge(ctx, cwd) {
 			sources: () => Promise.resolve(sourceSnapshot()),
 			saveSources: async (sources, expectedRevision) => {
 				const catalog = sources.filter((source) => !source.builtIn).map(validateCatalogSource);
-				if (new Set(catalog.map((source) => source.id)).size !== catalog.length) throw new Error("Catalog Source id 不能重复");
+				if (new Set(catalog.map((source) => source.id)).size !== catalog.length) throw new Error(ui("Catalog Source id 不能重复", "Catalog Source IDs must be unique"));
 				const reserved = new Set([NPM_SOURCE.id, ...(providers?.sources() ?? []).map((source) => source.id)]);
 				const conflict = catalog.find((source) => reserved.has(source.id));
-				if (conflict !== void 0) throw new Error(`Catalog Source ${conflict.id} 与内置或 Provider Source 冲突`);
+				if (conflict !== void 0) throw new Error(ui(`Catalog Source ${conflict.id} 与内置或 Provider Source 冲突`, `Catalog Source ${conflict.id} conflicts with a built-in or Provider Source`));
 				await mutateSettings(settings, MARKETPLACE_NAMESPACE, [{
 					op: "set",
 					path: ["sources"],
