+++ lib/marketplace-provider.js
@@ -1,4 +1,5 @@
-import { n as assertCredentialFreeUrl } from "./plugin-marketplace-D8-O8gcm.js";
+import { c as ui } from "./locale-DbTu1PjP.js";
+import { n as assertCredentialFreeUrl } from "./plugin-marketplace-swCofv4S.js";
 import { Service } from "@deepseek-ai/cordis";
 
 //#region src/host/marketplace-provider.ts
@@ -9,16 +10,16 @@ const RESERVED_SOURCE_IDS = new Set(["npm"]);
 const MAX_PROVIDER_RESULTS = 12;
 function optionalText(value, field, subject) {
 	if (value === void 0) return void 0;
-	if (typeof value !== "string") throw new TypeError(`${subject} 的 ${field} 必须是字符串`);
+	if (typeof value !== "string") throw new TypeError(ui(`${subject} 的 ${field} 必须是字符串`, `${field} of ${subject} must be a string`));
 	const text = value.trim();
 	return text === "" ? void 0 : text;
 }
 function normalizeSource(kind, source) {
-	if (!SOURCE_ID.test(source.id) || RESERVED_SOURCE_IDS.has(source.id)) throw new Error(`Source Provider ${kind} 的 Source id ${JSON.stringify(source.id)} 必须是非保留的小写 kebab-case`);
+	if (!SOURCE_ID.test(source.id) || RESERVED_SOURCE_IDS.has(source.id)) throw new Error(ui(`Source Provider ${kind} 的 Source id ${JSON.stringify(source.id)} 必须是非保留的小写 kebab-case`, `Source id ${JSON.stringify(source.id)} of Source Provider ${kind} must be a non-reserved lowercase kebab-case value`));
 	const label = source.label.trim();
 	const url = source.url.trim();
-	if (label === "" || url === "") throw new Error(`Source Provider ${kind} 的 Source 名称和 URL 不能为空`);
-	assertCredentialFreeUrl(url, `Source Provider ${kind} 的 URL`);
+	if (label === "" || url === "") throw new Error(ui(`Source Provider ${kind} 的 Source 名称和 URL 不能为空`, `Source name and URL of Source Provider ${kind} cannot be empty`));
+	assertCredentialFreeUrl(url, ui(`Source Provider ${kind} 的 URL`, `URL of Source Provider ${kind}`));
 	const credentialRef = optionalText(source.credentialRef, "credentialRef", `Source Provider ${kind}`);
 	return Object.freeze({
 		id: source.id,
@@ -32,8 +33,8 @@ function normalizeSource(kind, source) {
 }
 function normalizeDiscovery(value, kind) {
 	const spec = optionalText(value.spec, "spec", `Source Provider ${kind}`);
-	if (spec === void 0) throw new Error(`Source Provider ${kind} 返回了空插件 spec`);
-	assertCredentialFreeUrl(spec, `Source Provider ${kind} 的插件 spec`);
+	if (spec === void 0) throw new Error(ui(`Source Provider ${kind} 返回了空插件 spec`, `Source Provider ${kind} returned an empty plugin spec`));
+	assertCredentialFreeUrl(spec, ui(`Source Provider ${kind} 的插件 spec`, `plugin spec of Source Provider ${kind}`));
 	const name = optionalText(value.name, "name", `Source Provider ${kind}`);
 	const description = optionalText(value.description, "description", `Source Provider ${kind}`);
 	const publisher = optionalText(value.publisher, "publisher", `Source Provider ${kind}`);
@@ -58,14 +59,14 @@ var TuiMarketplaceProviderRegistry = class extends Service {
 	*/
 	register(provider) {
 		const kind = provider.kind.trim();
-		if (!PROVIDER_KIND.test(kind) || RESERVED_KINDS.has(kind)) throw new Error(`Source Provider kind ${JSON.stringify(provider.kind)} 必须是非保留的小写 kebab-case`);
-		if (this.providers.has(kind)) throw new Error(`Source Provider ${kind} 已注册`);
-		if (provider.sources.length === 0) throw new Error(`Source Provider ${kind} 必须至少声明一个 Source`);
+		if (!PROVIDER_KIND.test(kind) || RESERVED_KINDS.has(kind)) throw new Error(ui(`Source Provider kind ${JSON.stringify(provider.kind)} 必须是非保留的小写 kebab-case`, `Source Provider kind ${JSON.stringify(provider.kind)} must be a non-reserved lowercase kebab-case value`));
+		if (this.providers.has(kind)) throw new Error(ui(`Source Provider ${kind} 已注册`, `Source Provider ${kind} is already registered`));
+		if (provider.sources.length === 0) throw new Error(ui(`Source Provider ${kind} 必须至少声明一个 Source`, `Source Provider ${kind} must declare at least one Source`));
 		const sources = provider.sources.map((source) => normalizeSource(kind, source));
 		const ids = /* @__PURE__ */ new Set();
 		for (const source of sources) {
-			if (ids.has(source.id)) throw new Error(`Source Provider ${kind} 重复声明 Source ${source.id}`);
-			if (this.sourceIds.has(source.id)) throw new Error(`Marketplace Source ${source.id} 已由其他 Provider 注册`);
+			if (ids.has(source.id)) throw new Error(ui(`Source Provider ${kind} 重复声明 Source ${source.id}`, `Source Provider ${kind} declared Source ${source.id} more than once`));
+			if (this.sourceIds.has(source.id)) throw new Error(ui(`Marketplace Source ${source.id} 已由其他 Provider 注册`, `Marketplace Source ${source.id} is already registered by another Provider`));
 			ids.add(source.id);
 		}
 		const registered = Object.freeze({
@@ -98,8 +99,8 @@ var TuiMarketplaceProviderRegistry = class extends Service {
 	*/
 	async search(query, source, signal) {
 		const registered = this.providers.get(source.kind);
-		if (registered === void 0) throw new Error(`Source Provider ${JSON.stringify(source.kind)} 未注册或已卸载`);
-		if (!registered.sources.some((candidate) => candidate.id === source.id)) throw new Error(`Source ${JSON.stringify(source.id)} 不属于 Provider ${source.kind}`);
+		if (registered === void 0) throw new Error(ui(`Source Provider ${JSON.stringify(source.kind)} 未注册或已卸载`, `Source Provider ${JSON.stringify(source.kind)} is not registered or has been unloaded`));
+		if (!registered.sources.some((candidate) => candidate.id === source.id)) throw new Error(ui(`Source ${JSON.stringify(source.id)} 不属于 Provider ${source.kind}`, `Source ${JSON.stringify(source.id)} does not belong to Provider ${source.kind}`));
 		return (await registered.provider.search(query, source, signal)).slice(0, MAX_PROVIDER_RESULTS).map((result) => normalizeDiscovery(result, source.kind));
 	}
 };
