+++ lib/plugin-marketplace-swCofv4S.js
@@ -1,3 +1,4 @@
+import { c as ui } from "./locale-DbTu1PjP.js";
 import { isAbsolute, join, posix, relative, resolve, sep } from "node:path";
 import { existsSync, readFileSync, realpathSync, statSync } from "node:fs";
 import { fileURLToPath } from "node:url";
@@ -16,7 +17,9 @@ const MAX_PATCH_BYTES = 2 * 1024 * 1024;
 const MAX_TAR_ENTRIES = 8192;
 const MAX_SEARCH_RESULTS = 12;
 const DEFAULT_FETCH_TIMEOUT_MS = 15e3;
-const TARBALL_PENDING = "尚未下载 tarball；选中后才会校验 Bundle patch";
+function tarballPendingNotice() {
+	return ui("尚未下载 tarball；选中后才会校验 Bundle patch", "The tarball has not been downloaded yet; Bundle patch is checked after selection");
+}
 const SENSITIVE_QUERY_KEY = new RegExp(String.raw`(?:^|[-_])(?:access[-_]?token|api[-_]?key|auth|authorization|credential|password|secret|signature|token)(?:$|[-_])`, "i");
 function sourceType(spec) {
 	if (/^(?:git\+|github:|gitlab:|bitbucket:)|\.git(?:#|$)/i.test(spec)) return "git";
@@ -43,14 +46,14 @@ function assertCredentialFreeUrl(value, label) {
 	const raw = value.startsWith("git+") ? value.slice(4) : value;
 	if (!/^https?:\/\//i.test(raw)) return;
 	const url = new URL(raw);
-	if (url.username !== "" || url.password !== "") throw new Error(`${label} 不能在 URL 中内嵌用户名或 Secret；请使用 Credential Ref`);
+	if (url.username !== "" || url.password !== "") throw new Error(ui(`${label} 不能在 URL 中内嵌用户名或 Secret；请使用 Credential Ref`, `${label} must not embed a username or secret in the URL; use a Credential Ref`));
 	const sensitive = [...url.searchParams.keys()].find((key) => SENSITIVE_QUERY_KEY.test(key));
-	if (sensitive !== void 0) throw new Error(`${label} 不能在 URL query 中内嵌 ${JSON.stringify(sensitive)}；请使用 Credential Ref`);
+	if (sensitive !== void 0) throw new Error(ui(`${label} 不能在 URL query 中内嵌 ${JSON.stringify(sensitive)}；请使用 Credential Ref`, `${label} must not embed ${JSON.stringify(sensitive)} in the URL query; use a Credential Ref`));
 }
 function readLocalBounded(path, maxBytes, label) {
 	const stat = statSync(path);
-	if (!stat.isFile()) throw new Error(`${label} 不是普通文件`);
-	if (stat.size > maxBytes) throw new Error(`${label} 超过 ${maxBytes} 字节限制`);
+	if (!stat.isFile()) throw new Error(ui(`${label} 不是普通文件`, `${label} is not a regular file`));
+	if (stat.size > maxBytes) throw new Error(ui(`${label} 超过 ${maxBytes} 字节限制`, `${label} exceeds the ${maxBytes}-byte limit`));
 	return readFileSync(path);
 }
 function safePatchPath(patch) {
@@ -69,13 +72,13 @@ function parseJson(bytes, label) {
 	try {
 		return JSON.parse(new TextDecoder().decode(bytes));
 	} catch (error) {
-		throw new Error(`${label} 不是有效 JSON：${messageOf(error)}`);
+		throw new Error(ui(`${label} 不是有效 JSON：${messageOf(error)}`, `${label} is not valid JSON: ${messageOf(error)}`));
 	}
 }
 async function readBounded(response, maxBytes, label) {
-	if (!response.ok) throw new Error(`${label} 请求失败：HTTP ${response.status}`);
+	if (!response.ok) throw new Error(ui(`${label} 请求失败：HTTP ${response.status}`, `${label} request failed: HTTP ${response.status}`));
 	const length = Number(response.headers.get("content-length"));
-	if (Number.isFinite(length) && length > maxBytes) throw new Error(`${label} 超过 ${maxBytes} 字节限制`);
+	if (Number.isFinite(length) && length > maxBytes) throw new Error(ui(`${label} 超过 ${maxBytes} 字节限制`, `${label} exceeds the ${maxBytes}-byte limit`));
 	if (response.body === null) return new Uint8Array();
 	const reader = response.body.getReader();
 	const chunks = [];
@@ -87,7 +90,7 @@ async function readBounded(response, maxBytes, label) {
 			size += result.value.byteLength;
 			if (size > maxBytes) {
 				await reader.cancel();
-				throw new Error(`${label} 超过 ${maxBytes} 字节限制`);
+				throw new Error(ui(`${label} 超过 ${maxBytes} 字节限制`, `${label} exceeds the ${maxBytes}-byte limit`));
 			}
 			chunks.push(result.value);
 		}
@@ -107,7 +110,7 @@ function tarText(bytes, start, length) {
 }
 function tarSize(bytes, offset) {
 	const text = tarText(bytes, offset + 124, 12).trim();
-	if (!/^[0-7]*$/.test(text)) throw new Error("tarball 含无效文件大小");
+	if (!/^[0-7]*$/.test(text)) throw new Error(ui("tarball 含无效文件大小", "The tarball contains an invalid file size"));
 	return text === "" ? 0 : Number.parseInt(text, 8);
 }
 function paxPath(body) {
@@ -127,7 +130,7 @@ function paxPath(body) {
 }
 async function tarEntries(compressed) {
 	const bytes = compressed[0] === 31 && compressed[1] === 139 ? await gunzip$1(compressed, { maxOutputLength: MAX_INFLATED_BYTES }) : compressed;
-	if (bytes.byteLength > MAX_INFLATED_BYTES) throw new Error("tarball 解压后超过限制");
+	if (bytes.byteLength > MAX_INFLATED_BYTES) throw new Error(ui("tarball 解压后超过限制", "The inflated tarball exceeds the size limit"));
 	const entries = /* @__PURE__ */ new Map();
 	let offset = 0;
 	let count = 0;
@@ -135,11 +138,11 @@ async function tarEntries(compressed) {
 	while (offset + 512 <= bytes.byteLength) {
 		const header = bytes.subarray(offset, offset + 512);
 		if (header.every((byte) => byte === 0)) break;
-		if (++count > MAX_TAR_ENTRIES) throw new Error("tarball 文件数量超过限制");
+		if (++count > MAX_TAR_ENTRIES) throw new Error(ui("tarball 文件数量超过限制", "The tarball has too many files"));
 		const size = tarSize(bytes, offset);
 		const dataStart = offset + 512;
 		const dataEnd = dataStart + size;
-		if (!Number.isSafeInteger(dataEnd) || dataEnd > bytes.byteLength) throw new Error("tarball 文件边界无效");
+		if (!Number.isSafeInteger(dataEnd) || dataEnd > bytes.byteLength) throw new Error(ui("tarball 文件边界无效", "The tarball file boundary is invalid"));
 		const name = tarText(bytes, offset, 100);
 		const prefix = tarText(bytes, offset + 345, 155);
 		const type = String.fromCharCode(header[156] ?? 0);
@@ -161,13 +164,13 @@ function packageCandidate(manifest, patchBytes, facts) {
 	const patch = manifest.dsh?.bundle?.patch;
 	const normalized = patch === void 0 ? void 0 : safePatchPath(patch);
 	const patchValid = normalized !== void 0 && patchBytes !== void 0 && patchBytes.byteLength <= MAX_PATCH_BYTES && validPatchText(new TextDecoder().decode(patchBytes));
-	if (patch === void 0) diagnostics.push("package.json 未声明 dsh.bundle.patch");
-	else if (normalized === void 0) diagnostics.push("dsh.bundle.patch 必须是包内相对路径");
-	else if (patchBytes === void 0) diagnostics.push(`发布内容不含 ${normalized}`);
-	else if (patchBytes.byteLength > MAX_PATCH_BYTES) diagnostics.push(`${normalized} 超过 ${MAX_PATCH_BYTES} 字节限制`);
-	else if (!patchValid) diagnostics.push(`${normalized} 不是有效的 Loader patch 数组`);
+	if (patch === void 0) diagnostics.push(ui("package.json 未声明 dsh.bundle.patch", "package.json does not declare dsh.bundle.patch"));
+	else if (normalized === void 0) diagnostics.push(ui("dsh.bundle.patch 必须是包内相对路径", "dsh.bundle.patch must be a package-relative path"));
+	else if (patchBytes === void 0) diagnostics.push(ui(`发布内容不含 ${normalized}`, `The published contents do not include ${normalized}`));
+	else if (patchBytes.byteLength > MAX_PATCH_BYTES) diagnostics.push(ui(`${normalized} 超过 ${MAX_PATCH_BYTES} 字节限制`, `${normalized} exceeds the ${MAX_PATCH_BYTES}-byte limit`));
+	else if (!patchValid) diagnostics.push(ui(`${normalized} 不是有效的 Loader patch 数组`, `${normalized} is not a valid Loader patch array`));
 	const scripts = Object.keys(manifest.scripts ?? {});
-	if (scripts.length > 0) diagnostics.push(`安装包声明脚本：${scripts.join("、")}`);
+	if (scripts.length > 0) diagnostics.push(ui(`安装包声明脚本：${scripts.join("、")}`, `The package declares scripts: ${scripts.join(", ")}`));
 	return {
 		id: facts.id,
 		name: manifest.name ?? facts.spec,
@@ -187,9 +190,9 @@ function packageCandidate(manifest, patchBytes, facts) {
 async function manifestFromTarball(bytes, facts) {
 	const entries = await tarEntries(bytes);
 	const manifestBytes = entries.get("package/package.json") ?? entries.get("package.json");
-	if (manifestBytes === void 0) throw new Error("tarball 不含 package.json");
+	if (manifestBytes === void 0) throw new Error(ui("tarball 不含 package.json", "The tarball does not contain package.json"));
 	const parsed = parseJson(manifestBytes, "package.json");
-	if (typeof parsed !== "object" || parsed === null || typeof parsed.name !== "string") throw new Error("tarball package.json 缺少包名");
+	if (typeof parsed !== "object" || parsed === null || typeof parsed.name !== "string") throw new Error(ui("tarball package.json 缺少包名", "The tarball package.json has no package name"));
 	const manifest = parsed;
 	const patch = manifest.dsh?.bundle?.patch;
 	const normalized = patch === void 0 ? void 0 : safePatchPath(patch);
@@ -209,7 +212,7 @@ function pendingSearchCandidate(facts) {
 		patchValid: false,
 		scripts: [],
 		immutable: facts.immutable,
-		diagnostics: [TARBALL_PENDING]
+		diagnostics: [tarballPendingNotice()]
 	};
 }
 function parseNpmSpec(spec) {
@@ -217,7 +220,7 @@ function parseNpmSpec(spec) {
 	if (value.startsWith("@")) {
 		const slash = value.indexOf("/");
 		const at$1 = value.lastIndexOf("@");
-		if (slash <= 1) throw new Error(`无效 npm 包名 ${JSON.stringify(spec)}`);
+		if (slash <= 1) throw new Error(ui(`无效 npm 包名 ${JSON.stringify(spec)}`, `Invalid npm package name ${JSON.stringify(spec)}`));
 		return at$1 > slash ? {
 			name: value.slice(0, at$1),
 			version: value.slice(at$1 + 1)
@@ -258,7 +261,7 @@ var PluginMarketplace = class {
 	*/
 	async search(query, sources, signal) {
 		const text = query.trim();
-		if (text === "") throw new Error("插件搜索词不能为空");
+		if (text === "") throw new Error(ui("插件搜索词不能为空", "Plugin search query cannot be empty"));
 		const enabled = sources.filter((source) => source.enabled);
 		const cacheKey = `search:${text}:${enabled.map((source) => source.id).join(",")}`;
 		const cached = this.searchCache.get(cacheKey);
@@ -287,7 +290,7 @@ var PluginMarketplace = class {
 				patchValid: false,
 				scripts: [],
 				immutable: false,
-				diagnostics: [`来源失败：${message}`]
+				diagnostics: [ui(`来源失败：${message}`, `Source failed: ${message}`)]
 			});
 		}
 		const result = [...deduped.values()];
@@ -303,7 +306,7 @@ var PluginMarketplace = class {
 	*/
 	async inspect(spec, sources, signal) {
 		const value = spec.trim();
-		if (value === "") throw new Error("插件 spec 不能为空");
+		if (value === "") throw new Error(ui("插件 spec 不能为空", "Plugin spec cannot be empty"));
 		const cached = this.inspectCache.get(value);
 		if (cached !== void 0) return cached;
 		const candidate = await this.inspectUncached(value, sources, signal);
@@ -311,7 +314,7 @@ var PluginMarketplace = class {
 		return candidate;
 	}
 	async inspectUncached(value, sources, signal) {
-		assertCredentialFreeUrl(value, "插件 spec");
+		assertCredentialFreeUrl(value, ui("插件 spec", "plugin spec"));
 		const type = sourceType(value);
 		if (type === "git") {
 			const immutable = /#[0-9a-f]{7,40}$/i.test(value);
@@ -325,12 +328,12 @@ var PluginMarketplace = class {
 				patchValid: false,
 				scripts: [],
 				immutable,
-				diagnostics: ["Git 来源必须安装后由原生 Manager 再验证 Bundle；安装可能执行 prepare/install 脚本", ...immutable ? [] : ["Git spec 未固定 commit，后续内容可能变化"]]
+				diagnostics: [ui("Git 来源必须安装后由原生 Manager 再验证 Bundle；安装可能执行 prepare/install 脚本", "A Git source can be revalidated as a Bundle only after installation by the native Manager; installation may run prepare/install scripts"), ...immutable ? [] : [ui("Git spec 未固定 commit，后续内容可能变化", "The Git spec is not pinned to a commit; future content may change")]]
 			};
 		}
 		if (type === "local") return this.inspectLocal(value);
 		if (type === "tarball") {
-			if (/^https?:/i.test(value)) return manifestFromTarball(await readBounded(await this.fetcher(value, requestInit({}, this.timeoutMs, signal)), MAX_TARBALL_BYTES, "插件 tarball"), {
+			if (/^https?:/i.test(value)) return manifestFromTarball(await readBounded(await this.fetcher(value, requestInit({}, this.timeoutMs, signal)), MAX_TARBALL_BYTES, ui("插件 tarball", "plugin tarball")), {
 				id: `direct:${value}`,
 				sourceId: "direct",
 				source: "tarball",
@@ -340,7 +343,7 @@ var PluginMarketplace = class {
 			return this.inspectLocal(value);
 		}
 		const npm = sources.find((source) => source.kind === "npm" && source.enabled) ?? sources.find((source) => source.kind === "npm");
-		if (npm === void 0) throw new Error("没有可用 npm Registry Source");
+		if (npm === void 0) throw new Error(ui("没有可用 npm Registry Source", "No npm Registry Source is available"));
 		return this.inspectNpm(value, npm, signal);
 	}
 	async headers(source, target = source.url) {
@@ -355,7 +358,7 @@ var PluginMarketplace = class {
 		const url = new URL("-/v1/search", source.url.endsWith("/") ? source.url : `${source.url}/`);
 		url.searchParams.set("text", `${query} keywords:dsh`);
 		url.searchParams.set("size", String(MAX_SEARCH_RESULTS));
-		const body = parseJson(await readBounded(await this.fetcher(url, requestInit(await this.headers(source, url), this.timeoutMs, signal)), MAX_INDEX_BYTES, `npm Source ${source.label}`), "npm 搜索响应");
+		const body = parseJson(await readBounded(await this.fetcher(url, requestInit(await this.headers(source, url), this.timeoutMs, signal)), MAX_INDEX_BYTES, `npm Source ${source.label}`), ui("npm 搜索响应", "npm search response"));
 		return (typeof body === "object" && body !== null && Array.isArray(body.objects) ? body.objects : []).flatMap((row) => {
 			if (typeof row !== "object" || row === null) return [];
 			const pkg = row.package;
@@ -383,17 +386,17 @@ var PluginMarketplace = class {
 		const parsed = parseNpmSpec(spec);
 		const registry = source.url.endsWith("/") ? source.url : `${source.url}/`;
 		const metadataUrl = new URL(encodeURIComponent(parsed.name), registry);
-		const metadata = parseJson(await readBounded(await this.fetcher(metadataUrl, requestInit(await this.headers(source, metadataUrl), this.timeoutMs, signal)), MAX_INDEX_BYTES, `npm 包 ${parsed.name}`), "npm 包元数据");
-		if (typeof metadata !== "object" || metadata === null) throw new Error("npm 元数据不是对象");
+		const metadata = parseJson(await readBounded(await this.fetcher(metadataUrl, requestInit(await this.headers(source, metadataUrl), this.timeoutMs, signal)), MAX_INDEX_BYTES, ui(`npm 包 ${parsed.name}`, `npm package ${parsed.name}`)), ui("npm 包元数据", "npm package metadata"));
+		if (typeof metadata !== "object" || metadata === null) throw new Error(ui("npm 元数据不是对象", "npm metadata is not an object"));
 		const record = metadata;
 		const version = parsed.version ?? record["dist-tags"]?.latest;
-		if (version === void 0) throw new Error(`${parsed.name} 没有 latest 版本`);
+		if (version === void 0) throw new Error(ui(`${parsed.name} 没有 latest 版本`, `${parsed.name} has no latest version`));
 		const manifest = record.versions?.[version];
-		if (manifest === void 0) throw new Error(`${parsed.name} 不含版本 ${version}`);
+		if (manifest === void 0) throw new Error(ui(`${parsed.name} 不含版本 ${version}`, `${parsed.name} does not contain version ${version}`));
 		const tarball = manifest.dist?.tarball;
-		if (typeof tarball !== "string") throw new Error(`${parsed.name}@${version} 缺少 dist.tarball`);
+		if (typeof tarball !== "string") throw new Error(ui(`${parsed.name}@${version} 缺少 dist.tarball`, `${parsed.name}@${version} is missing dist.tarball`));
 		const tarballUrl = new URL(tarball);
-		if (!["http:", "https:"].includes(tarballUrl.protocol)) throw new Error("npm dist.tarball 必须使用 HTTP(S)");
+		if (!["http:", "https:"].includes(tarballUrl.protocol)) throw new Error(ui("npm dist.tarball 必须使用 HTTP(S)", "npm dist.tarball must use HTTP(S)"));
 		const tarResponse = await this.fetcher(tarballUrl, requestInit(await this.headers(source, tarballUrl), this.timeoutMs, signal));
 		const exactSpec = `${parsed.name}@${version}`;
 		const publisher = publisherOf(manifest);
@@ -405,7 +408,7 @@ var PluginMarketplace = class {
 			...publisher === void 0 ? {} : { publisher },
 			immutable: true
 		});
-		if (candidate.name !== parsed.name || candidate.version !== version) throw new Error(`${exactSpec} tarball 的 package.json 名称或版本与 Registry 元数据不一致`);
+		if (candidate.name !== parsed.name || candidate.version !== version) throw new Error(ui(`${exactSpec} tarball 的 package.json 名称或版本与 Registry 元数据不一致`, `The package.json name or version in the ${exactSpec} tarball does not match the Registry metadata`));
 		return {
 			...candidate,
 			...candidate.description === void 0 && manifest.description !== void 0 ? { description: manifest.description } : {}
@@ -416,11 +419,11 @@ var PluginMarketplace = class {
 		const prefix = /^(?:file:|link:)/.exec(spec)?.[0] ?? "";
 		const raw = prefix === "" ? spec : spec.slice(prefix.length);
 		const path = fileUrl ? fileURLToPath(spec) : resolve(this.cwd, raw);
-		if (!existsSync(path)) throw new Error(`本地插件不存在：${path}`);
+		if (!existsSync(path)) throw new Error(ui(`本地插件不存在：${path}`, `Local plugin does not exist: ${path}`));
 		const stat = statSync(path);
 		if (stat.isFile()) {
 			const finalSpec$1 = fileUrl ? spec : prefix === "" ? path : `${prefix}${path}`;
-			return await manifestFromTarball(readLocalBounded(path, MAX_TARBALL_BYTES, "本地插件 tarball"), {
+			return await manifestFromTarball(readLocalBounded(path, MAX_TARBALL_BYTES, ui("本地插件 tarball", "local plugin tarball")), {
 				id: `local:${path}`,
 				sourceId: "local",
 				source: "tarball",
@@ -428,10 +431,10 @@ var PluginMarketplace = class {
 				immutable: false
 			});
 		}
-		if (!stat.isDirectory()) throw new Error(`本地插件不是目录或 tarball：${path}`);
+		if (!stat.isDirectory()) throw new Error(ui(`本地插件不是目录或 tarball：${path}`, `Local plugin is not a directory or tarball: ${path}`));
 		const root = realpathSync(path);
 		const manifestPath = join(root, "package.json");
-		const manifest = JSON.parse(new TextDecoder().decode(readLocalBounded(manifestPath, MAX_MANIFEST_BYTES, "本地插件 package.json")));
+		const manifest = JSON.parse(new TextDecoder().decode(readLocalBounded(manifestPath, MAX_MANIFEST_BYTES, ui("本地插件 package.json", "local plugin package.json"))));
 		const patch = manifest.dsh?.bundle?.patch;
 		const normalized = patch === void 0 ? void 0 : safePatchPath(patch);
 		let patchBytes;
@@ -440,7 +443,7 @@ var PluginMarketplace = class {
 			if (existsSync(target)) {
 				const realTarget = realpathSync(target);
 				const within = relative(root, realTarget);
-				if (within !== ".." && !within.startsWith(`..${sep}`) && !isAbsolute(within)) patchBytes = readLocalBounded(realTarget, MAX_PATCH_BYTES, "本地 Bundle patch");
+				if (within !== ".." && !within.startsWith(`..${sep}`) && !isAbsolute(within)) patchBytes = readLocalBounded(realTarget, MAX_PATCH_BYTES, ui("本地 Bundle patch", "local Bundle patch"));
 			}
 		}
 		const finalSpec = fileUrl ? spec : prefix === "" ? path : `${prefix}${path}`;
@@ -480,7 +483,7 @@ var PluginMarketplace = class {
 		}));
 	}
 	async searchProvider(query, source, _sources, signal) {
-		if (this.providers === void 0) throw new Error(`Source Provider ${JSON.stringify(source.kind)} 未装配`);
+		if (this.providers === void 0) throw new Error(ui(`Source Provider ${JSON.stringify(source.kind)} 未装配`, `Source Provider ${JSON.stringify(source.kind)} is not mounted`));
 		return (await this.providers.search(query, source, signal)).map((entry) => pendingSearchCandidate({
 			id: `${source.id}:${entry.spec}`,
 			name: entry.name ?? entry.spec,
