+++ lib/startup.js
@@ -1,3 +1,4 @@
-import { i as name, n as apply, r as inject, t as TUI_STARTUP_SERVICE } from "./startup-DznulVSj.js";
+import "./locale-DbTu1PjP.js";
+import { i as name, n as apply, r as inject, t as TUI_STARTUP_SERVICE } from "./startup-Dk_6a6_Y.js";
 
 export { TUI_STARTUP_SERVICE, apply, inject, name };
