+++ src/host/marketplace-provider.ts
@@ -2,6 +2,7 @@
 
 import { Context, Service } from '@deepseek-ai/cordis'
 import type { TuiMarketplaceSource } from '@deepseek-ai/dsh-tui-protocol'
+import { ui } from '../client/locale.ts'
 import { assertCredentialFreeUrl } from './plugin-marketplace.ts'
 
 const PROVIDER_KIND = /^[a-z][a-z0-9-]*$/u
@@ -55,19 +56,32 @@ declare module '@deepseek-ai/cordis' {
 
 function optionalText(value: unknown, field: string, subject: string): string | undefined {
   if (value === undefined) return undefined
-  if (typeof value !== 'string') throw new TypeError(`${subject} 的 ${field} 必须是字符串`)
+  if (typeof value !== 'string') {
+    throw new TypeError(ui(
+      `${subject} 的 ${field} 必须是字符串`,
+      `${field} of ${subject} must be a string`,
+    ))
+  }
   const text = value.trim()
   return text === '' ? undefined : text
 }
 
 function normalizeSource(kind: string, source: TuiMarketplaceProviderSource): TuiMarketplaceSource {
   if (!SOURCE_ID.test(source.id) || RESERVED_SOURCE_IDS.has(source.id)) {
-    throw new Error(`Source Provider ${kind} 的 Source id ${JSON.stringify(source.id)} 必须是非保留的小写 kebab-case`)
+    throw new Error(ui(
+      `Source Provider ${kind} 的 Source id ${JSON.stringify(source.id)} 必须是非保留的小写 kebab-case`,
+      `Source id ${JSON.stringify(source.id)} of Source Provider ${kind} must be a non-reserved lowercase kebab-case value`,
+    ))
   }
   const label = source.label.trim()
   const url = source.url.trim()
-  if (label === '' || url === '') throw new Error(`Source Provider ${kind} 的 Source 名称和 URL 不能为空`)
-  assertCredentialFreeUrl(url, `Source Provider ${kind} 的 URL`)
+  if (label === '' || url === '') {
+    throw new Error(ui(
+      `Source Provider ${kind} 的 Source 名称和 URL 不能为空`,
+      `Source name and URL of Source Provider ${kind} cannot be empty`,
+    ))
+  }
+  assertCredentialFreeUrl(url, ui(`Source Provider ${kind} 的 URL`, `URL of Source Provider ${kind}`))
   const credentialRef = optionalText(source.credentialRef, 'credentialRef', `Source Provider ${kind}`)
   return Object.freeze({
     id: source.id,
@@ -82,8 +96,13 @@ function normalizeSource(kind: string, source: TuiMarketplaceProviderSource): Tu
 
 function normalizeDiscovery(value: TuiMarketplaceDiscovery, kind: string): TuiMarketplaceDiscovery {
   const spec = optionalText(value.spec, 'spec', `Source Provider ${kind}`)
-  if (spec === undefined) throw new Error(`Source Provider ${kind} 返回了空插件 spec`)
-  assertCredentialFreeUrl(spec, `Source Provider ${kind} 的插件 spec`)
+  if (spec === undefined) {
+    throw new Error(ui(
+      `Source Provider ${kind} 返回了空插件 spec`,
+      `Source Provider ${kind} returned an empty plugin spec`,
+    ))
+  }
+  assertCredentialFreeUrl(spec, ui(`Source Provider ${kind} 的插件 spec`, `plugin spec of Source Provider ${kind}`))
   const name = optionalText(value.name, 'name', `Source Provider ${kind}`)
   const description = optionalText(value.description, 'description', `Source Provider ${kind}`)
   const publisher = optionalText(value.publisher, 'publisher', `Source Provider ${kind}`)
@@ -112,17 +131,35 @@ export class TuiMarketplaceProviderRegistry extends Service {
   register(provider: TuiMarketplaceProvider): () => void {
     const kind = provider.kind.trim()
     if (!PROVIDER_KIND.test(kind) || RESERVED_KINDS.has(kind)) {
-      throw new Error(`Source Provider kind ${JSON.stringify(provider.kind)} 必须是非保留的小写 kebab-case`)
+      throw new Error(ui(
+        `Source Provider kind ${JSON.stringify(provider.kind)} 必须是非保留的小写 kebab-case`,
+        `Source Provider kind ${JSON.stringify(provider.kind)} must be a non-reserved lowercase kebab-case value`,
+      ))
     }
-    if (this.providers.has(kind)) throw new Error(`Source Provider ${kind} 已注册`)
+    if (this.providers.has(kind)) {
+      throw new Error(ui(`Source Provider ${kind} 已注册`, `Source Provider ${kind} is already registered`))
+    }
     if (provider.sources.length === 0) {
-      throw new Error(`Source Provider ${kind} 必须至少声明一个 Source`)
+      throw new Error(ui(
+        `Source Provider ${kind} 必须至少声明一个 Source`,
+        `Source Provider ${kind} must declare at least one Source`,
+      ))
     }
     const sources = provider.sources.map(source => normalizeSource(kind, source))
     const ids = new Set<string>()
     for (const source of sources) {
-      if (ids.has(source.id)) throw new Error(`Source Provider ${kind} 重复声明 Source ${source.id}`)
-      if (this.sourceIds.has(source.id)) throw new Error(`Marketplace Source ${source.id} 已由其他 Provider 注册`)
+      if (ids.has(source.id)) {
+        throw new Error(ui(
+          `Source Provider ${kind} 重复声明 Source ${source.id}`,
+          `Source Provider ${kind} declared Source ${source.id} more than once`,
+        ))
+      }
+      if (this.sourceIds.has(source.id)) {
+        throw new Error(ui(
+          `Marketplace Source ${source.id} 已由其他 Provider 注册`,
+          `Marketplace Source ${source.id} is already registered by another Provider`,
+        ))
+      }
       ids.add(source.id)
     }
     const registered = Object.freeze({ provider, sources: Object.freeze(sources) })
@@ -158,9 +195,17 @@ export class TuiMarketplaceProviderRegistry extends Service {
     signal?: AbortSignal,
   ): Promise<readonly TuiMarketplaceDiscovery[]> {
     const registered = this.providers.get(source.kind)
-    if (registered === undefined) throw new Error(`Source Provider ${JSON.stringify(source.kind)} 未注册或已卸载`)
+    if (registered === undefined) {
+      throw new Error(ui(
+        `Source Provider ${JSON.stringify(source.kind)} 未注册或已卸载`,
+        `Source Provider ${JSON.stringify(source.kind)} is not registered or has been unloaded`,
+      ))
+    }
     if (!registered.sources.some(candidate => candidate.id === source.id)) {
-      throw new Error(`Source ${JSON.stringify(source.id)} 不属于 Provider ${source.kind}`)
+      throw new Error(ui(
+        `Source ${JSON.stringify(source.id)} 不属于 Provider ${source.kind}`,
+        `Source ${JSON.stringify(source.id)} does not belong to Provider ${source.kind}`,
+      ))
     }
     const results = await registered.provider.search(query, source, signal)
     return results.slice(0, MAX_PROVIDER_RESULTS).map(result => normalizeDiscovery(result, source.kind))
